Add naming to config json, add time sync, change internet conn check
Use curl for internet connection check, since it seems
to be more reliable than urlopen. it's also what's used in the
dropbox script so it makes sense to use it here too

Add a dict of [serial] : [identifier] to the config json, this seemed
like a nice solution to identify where photos are coming from. This way
there's no need to change the user every time you re-flash the pi

Added in automatically synchronizing the rtc when the date it gives
back is before 2023.