Update state_update_tracker and get_files_to_upload
Both of these now check for error conditions and return accordingly
state_update_tracker creates an empty copy of its data to prevent errors
when the rest of the code attempts to access that data.

get_files_to_upload checks if the last_uploaded_file is in its folder of
files, otherwise it starts from the beginning. This is not ideal but we
don't currently have a better way. Keeping a database with all uploaded
files would be an improvement. If the file does not exist it returns the
list of files in images. if images is empty, it returns an empty list.

if the returned list is empty, the code does not attempt the upload, and
instead just updates its state and shuts down